requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohttp=3.8.5=pypi_0
aiosignal=1.3.1=pypi_0
arch=6.1.0=pypi_0
asttokens=2.2.1=pyhd8ed1ab_0
async-timeout=4.0.2=pypi_0
attrs=23.1.0=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
binance=0.3=pypi_0
bzip2=1.0.8=he774522_0
ca-certificates=2023.5.7=h56e8100_0
certifi=2023.7.22=pypi_0
charset-normalizer=3.2.0=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
dateparser=1.1.8=pypi_0
debugpy=1.5.1=py311hd77b12b_0
decorator=5.1.1=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
fastjsonschema=2.18.0=pypi_0
flatbuffers=23.5.26=pypi_0
frozenlist=1.4.0=pypi_0
greenlet=2.0.2=pypi_0
idna=3.4=pypi_0
importlib-metadata=6.8.0=pyha770c72_0
importlib_metadata=6.8.0=hd8ed1ab_0
ipykernel=6.15.0=pyh025b116_0
ipython=8.14.0=pyh08f2357_0
jedi=0.18.2=pyhd8ed1ab_0
joblib=1.3.1=pypi_0
jsonschema=4.18.4=pypi_0
jsonschema-specifications=2023.7.1=pypi_0
jupyter_client=8.3.0=pyhd8ed1ab_0
jupyter_core=5.3.1=py311h1ea47a8_0
libclang=16.0.6=pypi_0
libffi=3.4.4=hd77b12b_0
libsodium=1.0.18=h8d14728_1
matplotlib-inline=0.1.6=pyhd8ed1ab_0
multidict=6.0.4=pypi_0
nbformat=5.9.1=pypi_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
numpy=1.24.3=pypi_0
oauthlib=3.2.2=pypi_0
openssl=3.0.9=h2bbff1b_0
packaging=23.1=pyhd8ed1ab_0
pandas=2.0.3=pypi_0
parso=0.8.3=pyhd8ed1ab_0
patsy=0.5.3=pypi_0
pickleshare=0.7.5=py_1003
pip=23.1.2=py311haa95532_0
platformdirs=3.9.1=pyhd8ed1ab_0
plotly=5.15.0=pypi_0
prompt-toolkit=3.0.39=pyha770c72_0
prompt_toolkit=3.0.39=hd8ed1ab_0
protobuf=4.23.4=pypi_0
psutil=5.9.0=py311h2bbff1b_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyasn1=0.5.0=pypi_0
pycryptodome=3.18.0=pypi_0
pygments=2.15.1=pyhd8ed1ab_0
python=3.11.4=he1021f5_0
python-binance=1.0.17=pypi_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.11=2_cp311
pytz=2023.3=pypi_0
pywin32=305=py311h2bbff1b_0
pyzmq=25.1.0=py311hd77b12b_0
referencing=0.30.0=pypi_0
regex=2023.6.3=pypi_0
requests=2.31.0=pypi_0
rpds-py=0.9.2=pypi_0
scikit-learn=1.3.0=pypi_0
scipy=1.11.1=pypi_0
setuptools=67.8.0=py311haa95532_0
six=1.16.0=pyh6c4a22f_0
sqlalchemy=2.0.19=pypi_0
sqlite=3.41.2=h2bbff1b_0
stack_data=0.6.2=pyhd8ed1ab_0
statsmodels=0.14.0=pypi_0
ta=0.10.2=pypi_0
tenacity=8.2.2=pypi_0
tensorboard-data-server=0.7.1=pypi_0
tensorflow-estimator=2.13.0=pypi_0
tensorflow-io-gcs-filesystem=0.31.0=pypi_0
termcolor=2.3.0=pypi_0
threadpoolctl=3.2.0=pypi_0
tk=8.6.12=h2bbff1b_0
tornado=6.2=py311h2bbff1b_0
traitlets=5.9.0=pyhd8ed1ab_0
typing-extensions=4.5.0=pypi_0
tzdata=2023.3=pypi_0
tzlocal=5.0.1=pypi_0
ujson=5.8.0=pypi_0
urllib3=1.26.16=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.6=pyhd8ed1ab_0
websockets=11.0.3=pypi_0
wheel=0.38.4=py311haa95532_0
wrapt=1.15.0=pypi_0
xz=5.4.2=h8cc25b3_0
yarl=1.9.2=pypi_0
zeromq=4.3.4=h0e60522_1
zipp=3.16.2=pyhd8ed1ab_0
zlib=1.2.13=h8cc25b3_0
